Add Superb.podspec
Unfortunately CocoaPods really wants to generate its own umbrella header
for you, and even after disabling that, we have to go through some
contortions to get this to compile as expected in a CocoaPods environment.

This also lowers the deployment target to 8.0 (and 9.0 for SuperbDemo
which requires Safari View Controller).